move index.js to web/index.js
