+++ webpack/webpack.common.js
@@ -29,7 +29,6 @@ module.exports = {
             patterns: [
                 {from: 'web/favicon.png'},
                 {from: 'web/favicon-unread.png'},
-                {from: 'web/version.txt'},
             ],
         }),
     ],
